+++ packages/react-native-styled/src/stylesheet/context/StylesheetConsumer.tsx
@@ -1,6 +1,6 @@
 import { JSXElementConstructor, ReactNode } from "react";
 import StylesheetContext from "./StylesheetContext";
-import { AnyObject } from "@peersyst/react-types";
+import { AnyObject } from "@swisstype/essential";
 import { getComponentStylesheetName } from "../utils";
 
 export interface StylesheetConsumerProps<P extends { style?: AnyObject } = {}> {