+++ packages/react-native-styled/src/stylesheet/stylesheet.ts
@@ -6,7 +6,13 @@ import {
     StylesheetFunction,
     StylesheetWithMetadata,
 } from "../types";
-import { fromTheme, fromProps, fromDimensions, fromSafeAreaInsets } from "../accessors";
+import {
+    fromTheme,
+    fromProps,
+    fromDimensions,
+    fromSafeAreaInsets,
+    fromAppState,
+} from "../accessors";
 import { getComponentStylesheetName } from "./utils";
 
 /**
@@ -26,6 +32,7 @@ export default function stylesheet<P extends StyledComponentProps<P["style"]>>(
             fromProps,
             fromDimensions,
             fromSafeAreaInsets,
+            fromAppState,
         } as any) || {}) as Stylesheet<P["style"]> as StylesheetWithMetadata<P["style"]>;
         stylesheet._metadata = {
             component: getComponentStylesheetName(Component),